fix:product endpoint datas

+++ controllers/product.js
@@ -1,5 +1,4 @@
-const { Product, Airport } = require("../models");
-const airline = require("./airline");
+const { Product, Airport, Airplane, Airline } = require("../models");
 
 module.exports = {
   create: async (req, res, next) => {
@@ -60,41 +59,40 @@ module.exports = {
         });
       }
 
-      console.log(
-        airportFrom.name,
-        airportTo.name,
-        airplane.name,
-        airline.name,
-        airplane.capacity
-      );
-
       const product = await Product.create({
         iata_from,
         iata_to,
-        airport_from: airportFrom.name,
-        airport_to: airportTo.name,
         date_departure,
         date_arrival,
         est_time,
         price,
         gate,
         airplane_id,
-        airplane_name: airplane.name,
-        airline_name: airline.name,
         stock: airplane.capacity,
       });
+
+      let productDetail = {
+        airport_from: airportFrom.name,
+        airport_to: airportTo.name,
+        airplane: airplane.name,
+        airline: airline.name,
+      };
+
+      const result = { ...product.dataValues, ...productDetail };
 
       return res.status(200).json({
         status: true,
         message: "Product created",
-        data: product,
+        data: result,
       });
     } catch (err) {
       next(err);
     }
   },
+
   update: async (req, res, next) => {
     try {
+      const { id } = req.params;
       const {
         iata_from,
         iata_to,
@@ -104,6 +102,7 @@ module.exports = {
         price,
         gate,
         airplane_id,
+        stock,
       } = req.body;
 
       if (
@@ -151,34 +150,56 @@ module.exports = {
         });
       }
 
-      const product = await Product.update({
-        iata_from,
-        iata_to,
+      const product = await Product.update(
+        {
+          iata_from,
+          iata_to,
+          date_departure,
+          date_arrival,
+          est_time,
+          price,
+          gate,
+          airplane_id,
+          stock,
+        },
+        {
+          where: { id },
+        }
+      );
+
+      let productDetail = {
         airport_from: airportFrom.name,
         airport_to: airportTo.name,
-        date_departure,
-        date_arrival,
-        est_time,
-        price,
-        gate,
-        airplane_id,
-        airplane_name: airplane.name,
-        airline_name: airline.name,
-        stock: airplane.capacity,
-      });
+        airplane: airplane.name,
+        airline: airline.name,
+      };
+
+      const result = { ...req.body, ...productDetail };
 
       return res.status(200).json({
         status: true,
         message: "Product updated",
-        data: product,
+        data: result,
       });
     } catch (err) {
       next(err);
     }
   },
+
   delete: async (req, res, next) => {
     try {
       const { id } = req.params;
+
+      const productCheck = await Product.findOne({
+        where: { id },
+      });
+
+      if (!productCheck) {
+        return res.status(401).json({
+          status: false,
+          message: "Product not found",
+        });
+      }
 
       const product = await Product.destroy({
         where: { id },
@@ -187,12 +208,13 @@ module.exports = {
       return res.status(200).json({
         status: true,
         message: "Product deleted",
-        data: product,
+        data: { id },
       });
     } catch (err) {
       next(err);
     }
   },
+
   getAll: async (req, res, next) => {
     try {
       const product = await Product.findAll();
@@ -202,6 +224,33 @@ module.exports = {
           status: false,
           message: "Product not found",
         });
+      }
+
+      for (let i = 0; i < product.length; i++) {
+        const airportFrom = await Airport.findOne({
+          where: { iata: product[i].iata_from },
+        });
+        const airportTo = await Airport.findOne({
+          where: { iata: product[i].iata_to },
+        });
+
+        const airplane = await Airplane.findOne({
+          where: { id: product[i].airplane_id },
+        });
+
+        const airline = await Airline.findOne({
+          where: { id: airplane.airline_id },
+        });
+
+        let productDetail = {
+          airport_from: airportFrom.name,
+          airport_to: airportTo.name,
+          airplane: airplane.name,
+          airline: airline.name,
+        };
+
+        const result = { ...product[i].dataValues, ...productDetail };
+        product[i] = result;
       }
 
       return res.status(200).json({
@@ -228,10 +277,34 @@ module.exports = {
         });
       }
 
+      const airportFrom = await Airport.findOne({
+        where: { iata: product.iata_from },
+      });
+      const airportTo = await Airport.findOne({
+        where: { iata: product.iata_to },
+      });
+
+      const airplane = await Airplane.findOne({
+        where: { id: product.airplane_id },
+      });
+
+      const airline = await Airline.findOne({
+        where: { id: airplane.airline_id },
+      });
+
+      let productDetail = {
+        airport_from: airportFrom.name,
+        airport_to: airportTo.name,
+        airplane: airplane.name,
+        airline: airline.name,
+      };
+
+      const result = { ...product.dataValues, ...productDetail };
+
       return res.status(200).json({
         status: true,
         message: "Product found",
-        data: product,
+        data: result,
       });
     } catch (err) {
       next(err);
